organize code base for oauth pkg
