+++ project/spotipy/oauth/auth/base.py
@@ -1,4 +1,7 @@
-from spotipy.oauth import utils
+import abc
+import typing
+
+from spotipy.oauth import utils, cache
 
 
 class SpotifyBaseAuthenticator:
@@ -37,3 +40,55 @@ class SpotifyBaseAuthenticator:
     @property
     def authorize_url(self):
         return self.__oauth_authorize_url
+
+
+class SpotifyAuthFlow(typing.Protocol):
+
+    def get_access_token(self) -> str:
+        """
+        Retrieves an access token from the `Spotify API`
+        """
+
+
+class BaseAuthFlow(SpotifyBaseAuthenticator, abc.ABC):
+
+        # Including this here to
+        # signify this needful.
+        # all other `AuthFlow`s should
+        # be expected to have this attr.
+
+        def __init__(self, client_id: str, client_secret: str, *,
+            proxies: dict[str, str] = None,
+            session: utils.SpotifySession = None,
+            session_factory: utils.SessionFactory = None,
+            timeout: float | tuple[float, ...] = None,
+            cache_cls: type[cache.SpotifyCacheHandler] = None,
+            cache_params: dict[str, typing.Any] = None):
+
+            if not session:
+                session = utils.SpotifySession
+
+            super(BaseAuthFlow, self).__init__(session, session_factory=session_factory)
+
+            # Set the client_id and client_secret.
+            self.credentials.client_id     = client_id
+            self.credentials.client_secret = client_secret
+
+            # Apply additional values to the
+            # internal session.
+            self.session.proxies = proxies
+            self.session.verify  = True
+
+            # Homeless attributes.
+            self.timeout = timeout
+
+            if not cache_cls:
+                cache_cls = cache.FileCacheHandler
+            self.cache_cls    = cache_cls
+            self.cache_params = cache_params
+
+        @abc.abstractmethod
+        def get_access_token(self) -> str:
+            """
+            Retrieves an access token from the `Spotify API`
+            """
